Convert to URP + Fog plane implemented
Project converted to Universal Render Pipeline in order to take advantage of Shadergraph. Fog and outline shaders implemented for objects, although I will be needing to polish them a bit more. Due to issues with the outside wall outlines not displaying correctly on the phone screen, I used white cubes to simulate the outline effect to save my own sanity.